implement only axis 0
