+++ nusamai-voxelize/src/voxelize.rs
@@ -1,18 +1,11 @@
-use byteorder::{LittleEndian, WriteBytesExt};
 use nalgebra::{distance, Point3, Vector3};
-use nusamai_geometry::MultiPolygon;
-use serde_json::json;
-use std::{
-    collections::{HashMap, HashSet},
-    fs::File,
-    io::Write,
-};
+use std::collections::HashSet;
 
 #[derive(Eq, PartialEq, Hash, Debug)]
-struct Voxel {
-    x: i32,
-    y: i32,
-    z: i32,
+pub struct Voxel {
+    pub x: i32,
+    pub y: i32,
+    pub z: i32,
 }
 
 fn draw_line(voxels: &mut HashSet<Voxel>, start: [f64; 3], end: [f64; 3], voxel_size: f64) {
@@ -51,7 +44,7 @@ fn draw_line(voxels: &mut HashSet<Voxel>, start: [f64; 3], end: [f64; 3], voxel_
     }
 }
 
-fn triangle_to_voxel(triangles: &[[f64; 3]], voxel_size: f64) -> HashSet<Voxel> {
+pub fn triangle_to_voxel(triangles: &[[f64; 3]], voxel_size: f64) -> HashSet<Voxel> {
     // 占有されたボクセルを格納する
     // HashSetは重複を許さない
     let mut occupied_voxels = HashSet::new();
@@ -66,7 +59,7 @@ fn triangle_to_voxel(triangles: &[[f64; 3]], voxel_size: f64) -> HashSet<Voxel> 
     occupied_voxels
 }
 
-fn fill_triangle(voxels: &mut HashSet<Voxel>, voxel_size: f64, triangle: &[[f64; 3]]) {
+pub fn fill_triangle(voxels: &mut HashSet<Voxel>, voxel_size: f64, triangle: &[[f64; 3]]) {
     if triangle.len() != 3 {
         panic!("The number of vertices is not 3")
     }
@@ -75,18 +68,6 @@ fn fill_triangle(voxels: &mut HashSet<Voxel>, voxel_size: f64, triangle: &[[f64;
     let p1 = Point3::from(triangle[0]);
     let p2 = Point3::from(triangle[1]);
     let p3 = Point3::from(triangle[2]);
-    println!("p1={:?}, p2={:?}, p3={:?}", p1, p2, p3);
-
-    // bounding boxを算出
-    let mut min_point = p1.clone();
-    let mut max_point = p1.clone();
-    for p in &[p2, p3] {
-        for i in 0..3 {
-            min_point[i] = min_point[i].min(p[i]);
-            max_point[i] = max_point[i].max(p[i]);
-        }
-    }
-    let box_size = max_point - min_point;
 
     // 3辺の長さを算出し、三角形が小さい（すべての辺がvoxel_size未満）場合は、面を走査せずvoxelを一つだけ塗りつぶす
     if is_small_triangle(&p1, &p2, &p3, voxel_size) {
@@ -130,9 +111,6 @@ fn fill_triangle(voxels: &mut HashSet<Voxel>, voxel_size: f64, triangle: &[[f64;
     norm /= d;
 
     // 最大長の軸を求める
-    // norm_axis=0 (x) --> yz-plane
-    // norm_axis=1 (y) --> zx-plane
-    // norm_axis=2 (z) --> xy-plane
     let norm_axis = norm
         .abs()
         .iter()
@@ -141,33 +119,23 @@ fn fill_triangle(voxels: &mut HashSet<Voxel>, voxel_size: f64, triangle: &[[f64;
         .map(|(i, _)| i)
         .unwrap();
 
-    let sweep_axis = match norm_axis {
-        0 => {
-            if box_size[1] >= box_size[2] {
-                1
-            } else {
-                2
-            }
+    // 三角形のbounding boxを算出
+    let mut min_point = p1;
+    let mut max_point = p1;
+    for p in &[p2, p3] {
+        for i in 0..3 {
+            min_point[i] = min_point[i].min(p[i]);
+            max_point[i] = max_point[i].max(p[i]);
         }
-        1 => {
-            if box_size[2] >= box_size[0] {
-                2
-            } else {
-                0
-            }
-        }
-        _ => {
-            if box_size[0] >= box_size[1] {
-                0
-            } else {
-                1
-            }
-        }
-    };
+    }
+    let box_size = max_point - min_point;
+    println!("box_size: {:?}", box_size);
 
     // norm_axis=0 (x) --> yz-plane
     // norm_axis=1 (y) --> zx-plane
-    // norm_axis=2 (z) --> xy-plane
+    // norm_axis=3 (z) --> xy-plane
+
+    // 一番長い辺を見つける
     let sweep_axis = match norm_axis {
         0 => {
             if box_size[1] >= box_size[2] {
@@ -194,287 +162,202 @@ fn fill_triangle(voxels: &mut HashSet<Voxel>, voxel_size: f64, triangle: &[[f64;
 
     match sweep_axis {
         0 => {
-            let mut tri = triangle.to_vec();
-            tri.sort_by(|a, b| a[0].partial_cmp(&b[0]).unwrap());
-            assert!(tri[1][0] >= tri[0][0]);
+            let mut sorted_triangle = triangle.to_vec();
+            // 3点のx座標を比較していき、もっともx座標が小さいものが最初に、大きいものが最後になるようにソート
+            // つまり、頂点の順番が反時計回りになるようにソート
+            sorted_triangle.sort_by(|a, b| a[0].partial_cmp(&b[0]).unwrap());
+            assert!(sorted_triangle[1][0] >= sorted_triangle[0][0]);
+            println!(
+                "p0: {:?}, p1: {:?}, p2: {:?}",
+                sorted_triangle[0], sorted_triangle[1], sorted_triangle[2]
+            );
 
-            let (mut start, mut end, mut vend) = if (tri[1][0] - tri[0][0]).abs() > 0.0
-                && (tri[1][0] - tri[0][0].floor() > 1.0)
+            // d1・d2はここで求めておく必要がありそう
+            let edge_direction_1;
+
+            // 走査軸が0（X軸）でX座標を基準にソートされていることがわかっているのでその前提で処理を進めることができる
+            // 絶対値の差が0より大きい、つまり最初と2番目の頂点のx座標が異なる場合で、尚且つ最初の頂点と2番目の頂点のx座標の差がvoxel_sizeより大きい場合にTrueになる
+            let (mut start_point, mut end_point, mut end_vertex_x) = if (sorted_triangle[1][0]
+                - sorted_triangle[0][0])
+                .abs()
+                > 0.0
+                && (sorted_triangle[1][0] - sorted_triangle[0][0].floor() > voxel_size)
             {
-                let d1 = (Vector3::from(tri[1]) - Vector3::from(tri[0])) / (tri[1][0] - tri[0][0]);
+                edge_direction_1 = (Vector3::from(sorted_triangle[1])
+                    - Vector3::from(sorted_triangle[0]))
+                    / (sorted_triangle[1][0] - sorted_triangle[0][0]);
                 (
-                    Vector3::from(tri[0]) + d1 * (1.0 - tri[0][0] + tri[0][0].floor()),
-                    Vector3::from(tri[0])
-                        + (Vector3::from(tri[2]) - Vector3::from(tri[0])) / (tri[2][0] - tri[0][0])
-                            * (1.0 - tri[0][0] + tri[0][0].floor()),
-                    tri[1][0],
+                    Vector3::from(sorted_triangle[0])
+                        + edge_direction_1
+                            * (voxel_size - sorted_triangle[0][0] + sorted_triangle[0][0].floor()),
+                    Vector3::from(sorted_triangle[0])
+                        + (Vector3::from(sorted_triangle[2]) - Vector3::from(sorted_triangle[0]))
+                            / (sorted_triangle[2][0] - sorted_triangle[0][0])
+                            * (voxel_size - sorted_triangle[0][0] + sorted_triangle[0][0].floor()),
+                    sorted_triangle[1][0],
                 )
             } else {
-                let d1 = (Vector3::from(tri[2]) - Vector3::from(tri[1])) / (tri[2][0] - tri[1][0]);
+                // 頂点0と1が直線上に並んでいる、もしくは一直線ではないが、差がvoxel_sizeより小さく1つのボクセルに集約される場合の処理
+                edge_direction_1 = (Vector3::from(sorted_triangle[2])
+                    - Vector3::from(sorted_triangle[1]))
+                    / (sorted_triangle[2][0] - sorted_triangle[1][0]);
                 (
-                    Vector3::from(tri[1]) + d1 * (1.0 - tri[1][0] + tri[1][0].floor()),
+                    Vector3::from(sorted_triangle[1])
+                        + edge_direction_1
+                            * (voxel_size - sorted_triangle[1][0] + sorted_triangle[1][0].floor()),
+                    Vector3::from(sorted_triangle[0])
+                        + (Vector3::from(sorted_triangle[2]) - Vector3::from(sorted_triangle[0]))
+                            / (sorted_triangle[2][0] - sorted_triangle[0][0])
+                            * (voxel_size - sorted_triangle[0][0] + sorted_triangle[0][0].floor()),
+                    sorted_triangle[1][0],
+                )
+            };
+
+            let edge_direction_2 = (Vector3::from(sorted_triangle[2])
+                - Vector3::from(sorted_triangle[0]))
+                / (sorted_triangle[2][0] - sorted_triangle[0][0]);
+
+            println!("edge_direction_1: {:?}", edge_direction_1);
+            println!("edge_direction_2: {:?}", edge_direction_2);
+
+            println!("start_point: {:?}, end_point: {:?}", start_point, end_point);
+
+            if start_point.norm() > 1000.0 || end_point.norm() > 1000.0 {
+                println!("Direction vector magnitude is too large");
+                return;
+            }
+
+            while end_point[0] < sorted_triangle[2][0] {
+                println!("axis 0 start: {:?}, end: {:?}", start_point, end_point);
+                draw_line(voxels, start_point.into(), end_point.into(), voxel_size);
+
+                start_point += edge_direction_1;
+                end_point += edge_direction_2;
+
+                if start_point[0] >= end_vertex_x {
+                    end_vertex_x = start_point[0] - sorted_triangle[1][0];
+                    start_point -= edge_direction_1 * end_vertex_x;
+                    if (sorted_triangle[2][0] - sorted_triangle[1][0]).abs() < f64::EPSILON {
+                        break;
+                    }
+                    start_point += (Vector3::from(sorted_triangle[2])
+                        - Vector3::from(sorted_triangle[1]))
+                        / (sorted_triangle[2][0] - sorted_triangle[1][0])
+                        * end_vertex_x;
+                    end_vertex_x = sorted_triangle[2][0];
+                }
+            }
+        }
+        1 => {
+            let mut tri = triangle.to_vec();
+            tri.sort_by(|a, b| a[1].partial_cmp(&b[1]).unwrap());
+            assert!(tri[1][1] >= tri[0][1]);
+
+            let (mut start, mut end, mut vend) = if (tri[1][1] - tri[0][1]).abs() > 0.0
+                && (tri[1][1] - tri[0][1].floor() > 1.0)
+            {
+                let d1 = (Vector3::from(tri[1]) - Vector3::from(tri[0])) / (tri[1][1] - tri[0][1]);
+                (
+                    Vector3::from(tri[0]) + d1 * (1.0 - tri[0][1] + tri[0][1].floor()),
                     Vector3::from(tri[0])
-                        + (Vector3::from(tri[2]) - Vector3::from(tri[0])) / (tri[2][0] - tri[0][0])
-                            * (1.0 - tri[0][0] + tri[0][0].floor()),
-                    tri[1][0],
+                        + (Vector3::from(tri[2]) - Vector3::from(tri[0])) / (tri[2][1] - tri[0][1])
+                            * (1.0 - tri[0][1] + tri[0][1].floor()),
+                    tri[1][1],
+                )
+            } else {
+                let d1 = (Vector3::from(tri[2]) - Vector3::from(tri[1])) / (tri[2][1] - tri[1][1]);
+                (
+                    Vector3::from(tri[1]) + d1 * (1.0 - tri[1][1] + tri[1][1].floor()),
+                    Vector3::from(tri[0])
+                        + (Vector3::from(tri[2]) - Vector3::from(tri[0])) / (tri[2][1] - tri[0][1])
+                            * (1.0 - tri[0][1] + tri[0][1].floor()),
+                    tri[1][1],
                 )
             };
 
             if start.norm() > 1000.0 || end.norm() > 1000.0 {
-                println!("d warn");
+                println!("Direction vector magnitude is too large");
                 return;
             }
 
-            while end[0] < tri[2][0] {
-                println!("{:?}, {:?}", start, end);
-                // todo
+            while end[1] < tri[2][1] {
                 draw_line(voxels, start.into(), end.into(), voxel_size);
 
-                let d1 = (Vector3::from(tri[1]) - Vector3::from(tri[0])) / (tri[1][0] - tri[0][0]);
-                let d2 = (Vector3::from(tri[2]) - Vector3::from(tri[0])) / (tri[2][0] - tri[0][0]);
+                let d1 = (Vector3::from(tri[1]) - Vector3::from(tri[0])) / (tri[1][1] - tri[0][1]);
+                let d2 = (Vector3::from(tri[2]) - Vector3::from(tri[0])) / (tri[2][1] - tri[0][1]);
                 start += d1;
                 end += d2;
-                if start[0] >= vend {
-                    vend = start[0] - tri[1][0];
+                if start[1] >= vend {
+                    vend = start[1] - tri[1][1];
                     start -= d1 * vend;
-                    if (tri[2][0] - tri[1][0]).abs() < f64::EPSILON {
+                    if (tri[2][1] - tri[1][1]).abs() < f64::EPSILON {
                         break;
                     }
                     start += (Vector3::from(tri[2]) - Vector3::from(tri[1]))
-                        / (tri[2][0] - tri[1][0])
+                        / (tri[2][1] - tri[1][1])
                         * vend;
-                    vend = tri[2][0];
+                    vend = tri[2][1];
                 }
             }
         }
-        1 => {}
-        2 => {}
-        _ => {}
+        2 => {
+            let mut tri = triangle.to_vec();
+            tri.sort_by(|a, b| a[2].partial_cmp(&b[2]).unwrap());
+            assert!(tri[1][2] >= tri[0][2]);
+
+            let (mut start, mut end, mut vend) = if (tri[1][2] - tri[0][2]).abs() > 1e-1 {
+                let d1 = (Vector3::from(tri[1]) - Vector3::from(tri[0])) / (tri[1][2] - tri[0][2]);
+                (
+                    Vector3::from(tri[0]) + d1 * (1.0 - tri[0][2] + tri[0][2].floor()),
+                    Vector3::from(tri[0])
+                        + (Vector3::from(tri[2]) - Vector3::from(tri[0])) / (tri[2][2] - tri[0][2])
+                            * (1.0 - tri[0][2] + tri[0][2].floor()),
+                    tri[1][2],
+                )
+            } else {
+                let d1 = (Vector3::from(tri[2]) - Vector3::from(tri[1])) / (tri[2][2] - tri[1][2]);
+                (
+                    Vector3::from(tri[1]) + d1 * (1.0 - tri[1][2] + tri[1][2].floor()),
+                    Vector3::from(tri[0])
+                        + (Vector3::from(tri[2]) - Vector3::from(tri[0])) / (tri[2][2] - tri[0][2])
+                            * (1.0 - tri[0][2] + tri[0][2].floor()),
+                    tri[1][2],
+                )
+            };
+
+            if start.norm() > 1000.0 || end.norm() > 1000.0 {
+                println!("Direction vector magnitude is too large");
+                return;
+            }
+
+            while end[2] < tri[2][2] {
+                draw_line(voxels, start.into(), end.into(), voxel_size);
+
+                let d1 = (Vector3::from(tri[1]) - Vector3::from(tri[0])) / (tri[1][2] - tri[0][2]);
+                let d2 = (Vector3::from(tri[2]) - Vector3::from(tri[0])) / (tri[2][2] - tri[0][2]);
+                start += d1;
+                end += d2;
+                if start[2] >= vend {
+                    vend = start[2] - tri[1][2];
+                    start -= d1 * vend;
+                    if (tri[2][2] - tri[1][2]).abs() < f64::EPSILON {
+                        break;
+                    }
+                    start += (Vector3::from(tri[2]) - Vector3::from(tri[1]))
+                        / (tri[2][2] - tri[1][2])
+                        * vend;
+                    vend = tri[2][2];
+                }
+            }
+        }
+        _ => unreachable!(),
     }
 }
 
-fn is_small_triangle(p1: &Point3<f64>, p2: &Point3<f64>, p3: &Point3<f64>, size: f64) -> bool {
+pub fn is_small_triangle(p1: &Point3<f64>, p2: &Point3<f64>, p3: &Point3<f64>, size: f64) -> bool {
     let d12 = distance(p1, p2);
     let d23 = distance(p2, p3);
     let d31 = distance(p3, p1);
 
     d12 <= size && d23 <= size && d31 <= size
-}
-
-#[cfg(test)]
-mod tests {
-    use earcut::{utils3d::project3d_to_2d, Earcut};
-
-    use super::*;
-
-    #[test]
-    fn test_voxelize() {
-        let vertices: Vec<[f64; 3]> = vec![
-            // Large box Exterior
-            [0.0, 0.0, 0.0],
-            [10.0, 0.0, 0.0],
-            [10.0, 10.0, 0.0],
-            [0.0, 10.0, 0.0],
-            [0.0, 0.0, 10.0],
-            [10.0, 0.0, 10.0],
-            [10.0, 10.0, 10.0],
-            [0.0, 10.0, 10.0],
-            // Large box Interior
-            [1.0, 1.0, 1.0],
-            [9.0, 1.0, 1.0],
-            [9.0, 9.0, 1.0],
-            [1.0, 9.0, 1.0],
-            [1.0, 1.0, 9.0],
-            [9.0, 1.0, 9.0],
-            [9.0, 9.0, 9.0],
-            [1.0, 9.0, 9.0],
-            // Medium box Exterior
-            [15.0, 0.0, 0.0],
-            [20.0, 0.0, 0.0],
-            [20.0, 5.0, 0.0],
-            [15.0, 5.0, 0.0],
-            [15.0, 0.0, 5.0],
-            [20.0, 0.0, 5.0],
-            [20.0, 5.0, 5.0],
-            [15.0, 5.0, 5.0],
-            // Medium box Interior
-            [16.0, 1.0, 1.0],
-            [19.0, 1.0, 1.0],
-            [19.0, 4.0, 1.0],
-            [16.0, 4.0, 1.0],
-            [16.0, 1.0, 4.0],
-            [19.0, 1.0, 4.0],
-            [19.0, 4.0, 4.0],
-            [16.0, 4.0, 4.0],
-            // Small box Exterior
-            [25.0, 0.0, 0.0],
-            [28.0, 0.0, 0.0],
-            [28.0, 3.0, 0.0],
-            [25.0, 3.0, 0.0],
-            [25.0, 0.0, 3.0],
-            [28.0, 0.0, 3.0],
-            [28.0, 3.0, 3.0],
-            [25.0, 3.0, 3.0],
-        ];
-
-        let mut mpoly = MultiPolygon::<u32>::new();
-
-        // Large box
-        mpoly.add_exterior([0, 1, 2, 3, 0]);
-        mpoly.add_exterior([4, 5, 6, 7, 4]);
-        mpoly.add_exterior([0, 1, 5, 4, 0]);
-        mpoly.add_exterior([1, 2, 6, 5, 1]);
-        mpoly.add_exterior([2, 3, 7, 6, 2]);
-        mpoly.add_exterior([3, 0, 4, 7, 3]);
-        mpoly.add_interior([8, 9, 10, 11, 8]);
-        mpoly.add_interior([12, 13, 14, 15, 12]);
-        mpoly.add_interior([8, 9, 13, 12, 8]);
-        mpoly.add_interior([9, 10, 14, 13, 9]);
-        mpoly.add_interior([10, 11, 15, 14, 10]);
-        mpoly.add_interior([11, 8, 12, 15, 11]);
-
-        // Medium box
-        mpoly.add_exterior([16, 17, 18, 19, 16]);
-        mpoly.add_exterior([20, 21, 22, 23, 20]);
-        mpoly.add_exterior([16, 17, 21, 20, 16]);
-        mpoly.add_exterior([17, 18, 22, 21, 17]);
-        mpoly.add_exterior([18, 19, 23, 22, 18]);
-        mpoly.add_exterior([19, 16, 20, 23, 19]);
-        mpoly.add_interior([24, 25, 26, 27, 24]);
-        mpoly.add_interior([28, 29, 30, 31, 28]);
-        mpoly.add_interior([24, 25, 29, 28, 24]);
-        mpoly.add_interior([25, 26, 30, 29, 25]);
-        mpoly.add_interior([26, 27, 31, 30, 26]);
-        mpoly.add_interior([27, 24, 28, 31, 27]);
-
-        // Small box
-        mpoly.add_exterior([32, 33, 34, 35, 32]);
-        mpoly.add_exterior([36, 37, 38, 39, 36]);
-        mpoly.add_exterior([32, 33, 37, 36, 32]);
-        mpoly.add_exterior([33, 34, 38, 37, 33]);
-        mpoly.add_exterior([34, 35, 39, 38, 34]);
-        mpoly.add_exterior([35, 32, 36, 39, 35]);
-
-        // triangulation
-        let mut earcutter = Earcut::new();
-        let mut buf3d: Vec<[f64; 3]> = Vec::new();
-        let mut buf2d: Vec<[f64; 2]> = Vec::new();
-        let mut index_buf: Vec<u32> = Vec::new();
-
-        let mut triangles: Vec<[f64; 3]> = Vec::new();
-
-        // ポリゴンを取り出す
-        for idx_poly in mpoly.iter() {
-            // インデックスを座標に変換
-            let poly = idx_poly.transform(|idx| vertices[*idx as usize]);
-            // holeがあるか確認
-            let num_outer = match poly.hole_indices().first() {
-                Some(&v) => v as usize,
-                None => poly.raw_coords().len(),
-            };
-
-            // 3次元での座標を格納
-            buf3d.clear();
-            buf3d.extend(poly.raw_coords().iter());
-
-            // 3次元座標を2次元座標に変換
-            if project3d_to_2d(&buf3d, num_outer, &mut buf2d) {
-                // earcut
-                earcutter.earcut(buf2d.iter().cloned(), poly.hole_indices(), &mut index_buf);
-                triangles.extend(index_buf.iter().map(|&idx| buf3d[idx as usize]));
-            }
-        }
-
-        let voxel_size = 1.0;
-
-        let occupied_voxels = triangle_to_voxel(&triangles, voxel_size);
-        let points_count = occupied_voxels.len();
-
-        // -------------------gltfの作成-------------------
-
-        // voxelは整数値だが、accessorsのcomponentTypeは5126（浮動小数点数）であり、primitivesの制約でINTEGER型は使用できない
-        let mut min_point = [f32::MAX; 3];
-        let mut max_point = [f32::MIN; 3];
-
-        let mut bin_file = File::create("data/output.bin").unwrap();
-        for v in occupied_voxels.iter() {
-            let [x, y, z] = [
-                (v.x as f32) * voxel_size as f32,
-                (v.y as f32) * voxel_size as f32,
-                (v.z as f32) * voxel_size as f32,
-            ];
-            min_point = [
-                f32::min(min_point[0], x),
-                f32::min(min_point[1], y),
-                f32::min(min_point[2], z),
-            ];
-            max_point = [
-                f32::max(max_point[0], x),
-                f32::max(max_point[1], y),
-                f32::max(max_point[2], z),
-            ];
-
-            bin_file.write_f32::<LittleEndian>(x).unwrap();
-            bin_file.write_f32::<LittleEndian>(y).unwrap();
-            bin_file.write_f32::<LittleEndian>(z).unwrap();
-        }
-
-        // voxelの数 × 頂点の座標数（3） × 4バイト（f32）
-        let byte_length = points_count * 3 * 4;
-
-        // GLTFファイルの作成
-        let gltf_json = json!( {
-            "asset": {
-                "version": "2.0",
-            },
-            "scene": 0,
-            "scenes": [
-                {
-                    "nodes": [0],
-                },
-            ],
-            "nodes": [
-                {"mesh": 0},
-            ],
-            "meshes": [
-                {
-                    "primitives": [
-                        {
-                            "attributes": {"POSITION": 0},
-                            "mode": 0,
-                        },
-                    ],
-                },
-            ],
-            "buffers": [
-                {
-                    "uri": "./output.bin",
-                    "byteLength": byte_length,
-                },
-            ],
-            "bufferViews": [
-                {
-                    "buffer": 0,
-                    "byteOffset": 0,
-                    "byteLength": byte_length,
-                    "target": 34962,
-                },
-            ],
-            "accessors": [
-                {
-                    "bufferView": 0,
-                    "byteOffset": 0,
-                    "componentType": 5126,
-                    "count": points_count,
-                    "type": "VEC3",
-                    "min": [min_point[0], min_point[1], min_point[2]],
-                    "max": [max_point[0], max_point[1], max_point[2]],
-                },
-            ],
-        });
-
-        // gltfファイルを出力
-        let mut gltf_file = File::create("data/output.gltf").unwrap();
-        let _ = gltf_file.write_all(gltf_json.to_string().as_bytes());
-    }
 }
